Brands Module done

+++ app/Http/Controllers/admin/BrandController.php
@@ -21,7 +21,7 @@ class BrandController extends Controller
     {
         $sort_search =null;
         $brands = Brand::orderBy('name', 'asc');
-        if ($request->has('search')){
+        if ($request->has('search') && $request->search != null){
             $sort_search = $request->search;
             $brands = $brands->where('name', 'like', '%'.$sort_search.'%');
         }
@@ -36,6 +36,7 @@ class BrandController extends Controller
      */
     public function create()
     {
+        return view('admin.brands.create');
     }
 
     /**
@@ -50,22 +51,18 @@ class BrandController extends Controller
         $brand->name = $request->name;
         $brand->meta_title = $request->meta_title;
         $brand->meta_description = $request->meta_description;
-        if ($request->slug != null) {
-            $brand->slug = str_replace(' ', '-', $request->slug);
+        if (Brand::where('slug',Str::slug($request->name))->count() > 0) {
+            $brand->slug = Str::slug($request->name).'-'.Str::random(5);
+        }else{
+            $brand->slug = Str::slug($request->name);
         }
-        else {
-            $brand->slug = preg_replace('/[^A-Za-z0-9\-]/', '', str_replace(' ', '-', $request->name)).'-'.Str::random(5);
-        }
-
         $brand->logo = $request->logo;
         $brand->save();
-
         $brand_translation = BrandTranslation::firstOrNew(['lang' => env('DEFAULT_LANGUAGE'), 'brand_id' => $brand->id]);
         $brand_translation->name = $request->name;
         $brand_translation->save();
 
-        // flash(translate('Brand has been inserted successfully'))->success();
-        return redirect()->route('admin.brands.index');
+        return redirect()->route('admin.brands.index')->with('success',translate('Brand has been inserted successfully'));
 
     }
 
@@ -103,17 +100,16 @@ class BrandController extends Controller
     public function update(Request $request, $id)
     {
         $brand = Brand::findOrFail($id);
+        $request->validate([
+            'slug'         => 'unique:brands,slug,'.$brand->id,
+        ]);
+
         if($request->lang == env("DEFAULT_LANGUAGE")){
             $brand->name = $request->name;
+            $brand->slug = $request->slug ?? Str::slug($request->name);
         }
         $brand->meta_title = $request->meta_title;
         $brand->meta_description = $request->meta_description;
-        if ($request->slug != null) {
-            $brand->slug = strtolower($request->slug);
-        }
-        else {
-            $brand->slug = preg_replace('/[^A-Za-z0-9\-]/', '', str_replace(' ', '-', $request->name)).'-'.Str::random(5);
-        }
         $brand->logo = $request->logo;
         $brand->save();
 
@@ -121,9 +117,7 @@ class BrandController extends Controller
         $brand_translation->name = $request->name;
         $brand_translation->save();
 
-        // flash(translate('Brand has been updated successfully'))->success();
-        return back();
-
+        return back()->with('success',translate('Brand has been updated successfully'));
     }
 
     /**
@@ -134,15 +128,25 @@ class BrandController extends Controller
      */
     public function destroy($id)
     {
-        $brand = Brand::findOrFail($id);
-        Product::where('brand_id', $brand->id)->delete();
-        foreach ($brand->brand_translations as $key => $brand_translation) {
-            $brand_translation->delete();
+        $brand = Brand::where('_id',$id)->first();
+        if (Product::where('brand_id', $brand->id)->count() > 0) {
+            return redirect()->back()->with('warning',translate('Brand can`t not be deleted.Products exist in this brand'));
+        }
+        if ($brand->brand_translations->count() > 0) {
+            foreach ($brand->brand_translations as $key => $brand_translation) {
+                $brand_translation->delete();
+            }
         }
         Brand::destroy($id);
+        return redirect()->route('admin.brands.index')->with('danger',translate('Brand has been deleted successfully'));
+    }
 
-        // flash(translate('Brand has been deleted successfully'))->success();
-        return redirect()->route('brands.index');
-
+    public function bulk_brand_delete(Request $request) {
+        if($request->id) {
+            foreach ($request->id as $brand_id) {
+                $this->destroy($brand_id);
+            }
+        }
+        return 1;
     }
 }
